refactor: create review object
