+++ internal/dto/leasing_history_body.go
@@ -15,9 +15,3 @@ type LeasingHistory struct {
 	End    time.Time           `json:"end"`
 	Price  float64             `json:"price"`
 }
-
-type Review struct {
-	Message  string    `json:"message"`
-	Rate     int       `json:"rate"`
-	CreateAt time.Time `json:"crateAt"`
-}
